test(chain): share loopback JSON-RPC harness
