+++ packages/chain/test/strict-current-finalized-evm-rpc.unit.test.ts
@@ -1,6 +1,3 @@
-import { createServer, type IncomingMessage, type Server, type ServerResponse } from 'node:http';
-import type { AddressInfo } from 'node:net';
-
 import {
   type ChainIdV1,
   type EvmAddressV1,
@@ -38,6 +35,12 @@ import {
   readStrictCurrentFinalizedEvmRevertDataV1,
   type StrictCurrentFinalizedEvmRpcConfigV1,
 } from '../src/strict-current-finalized-evm-rpc.js';
+import {
+  createLoopbackJsonRpcTestHarness,
+  sendJsonRpcError as sendError,
+  sendJsonRpcResult as sendResult,
+  type LoopbackJsonRpcHandler as LoopbackHandler,
+} from './loopback-json-rpc-test-helpers.js';
 
 const CHAIN_ID = '20430' as ChainIdV1;
 const CHAIN_QUANTITY = '0x4fce';
@@ -52,29 +55,11 @@ const WRONG_MAGIC_RETURN = `0xffffffff${'00'.repeat(28)}`;
 const FIRST_READ_DATA = '0x11111111';
 const SECOND_READ_DATA = '0x22222222';
 
-interface JsonRpcRequest {
-  readonly jsonrpc: '2.0';
-  readonly id: number;
-  readonly method: string;
-  readonly params: readonly unknown[];
-}
-
-interface LoopbackRpcServer {
-  readonly url: string;
-  readonly calls: JsonRpcRequest[];
-  readonly stop: () => Promise<void>;
-}
-
-type LoopbackHandler = (
-  call: JsonRpcRequest,
-  response: ServerResponse,
-  request: IncomingMessage,
-) => void | Promise<void>;
-
-const activeServers: LoopbackRpcServer[] = [];
+const rpcHarness = createLoopbackJsonRpcTestHarness();
+const startRpcServer = rpcHarness.start;
 
 afterEach(async () => {
-  await Promise.all(activeServers.splice(0).map((server) => server.stop()));
+  await rpcHarness.stopAll();
 });
 
 describe('RFC-64 strict current-finalized raw JSON-RPC transport', () => {
@@ -844,74 +829,6 @@ function successfulHandler(options: SuccessfulHandlerOptions = {}): LoopbackHand
         sendError(response, call, -32601, 'method not found');
     }
   };
-}
-
-async function startRpcServer(handler: LoopbackHandler): Promise<LoopbackRpcServer> {
-  const calls: JsonRpcRequest[] = [];
-  const server = createServer(async (request, response) => {
-    try {
-      const body = await readRequestBody(request);
-      const parsed = JSON.parse(body) as JsonRpcRequest;
-      calls.push(parsed);
-      await handler(parsed, response, request);
-    } catch (cause) {
-      if (!response.headersSent) response.writeHead(500, { 'content-type': 'text/plain' });
-      if (!response.writableEnded) response.end(cause instanceof Error ? cause.message : 'failure');
-    }
-  });
-  await new Promise<void>((resolve, reject) => {
-    server.once('error', reject);
-    server.listen(0, '127.0.0.1', () => {
-      server.off('error', reject);
-      resolve();
-    });
-  });
-  const address = server.address() as AddressInfo;
-  let stopped = false;
-  const loopback = Object.freeze({
-    url: `http://127.0.0.1:${address.port}`,
-    calls,
-    stop: async () => {
-      if (stopped) return;
-      stopped = true;
-      await closeServer(server);
-    },
-  });
-  activeServers.push(loopback);
-  return loopback;
-}
-
-async function readRequestBody(request: IncomingMessage): Promise<string> {
-  const chunks: Buffer[] = [];
-  for await (const chunk of request) chunks.push(Buffer.from(chunk));
-  return Buffer.concat(chunks).toString('utf8');
-}
-
-function sendResult(response: ServerResponse, request: JsonRpcRequest, result: unknown): void {
-  response.writeHead(200, { 'content-type': 'application/json' });
-  response.end(JSON.stringify({ jsonrpc: '2.0', id: request.id, result }));
-}
-
-function sendError(
-  response: ServerResponse,
-  request: JsonRpcRequest,
-  code: number,
-  message: string,
-  data?: string,
-): void {
-  response.writeHead(200, { 'content-type': 'application/json' });
-  response.end(JSON.stringify({
-    jsonrpc: '2.0',
-    id: request.id,
-    error: { code, message, ...(data === undefined ? {} : { data }) },
-  }));
-}
-
-async function closeServer(server: Server): Promise<void> {
-  await new Promise<void>((resolve) => {
-    server.close(() => resolve());
-    server.closeAllConnections();
-  });
 }
 
 function fixedRequest(
